Refactor platform detection and registry
